+++ syndicate/core/generators/project.py
@@ -29,7 +29,7 @@ from syndicate.core.generators.contents import (_get_lambda_default_policy,
                                                 GITIGNORE_CONTENT,
                                                 README_TEMPLATE)
 from syndicate.core.groups import (RUNTIME_JAVA, RUNTIME_NODEJS,
-                                   RUNTIME_PYTHON)
+                                   RUNTIME_PYTHON, RUNTIME_DOTNET)
 
 _LOG = get_logger('syndicate.core.generators.project')
 
@@ -103,8 +103,13 @@ def _generate_nodejs_project_hierarchy(full_project_path, project_name=None):
     _mkdir(os.path.join(full_project_path, FOLDER_COMMONS), exist_ok=True)
 
 
+def _generate_dotnet_project_hierarchy(full_project_path, project_name=None):
+    _mkdir(os.path.join(full_project_path, FOLDER_LAMBDAS), exist_ok=True)
+
+
 PROJECT_PROCESSORS = {
     RUNTIME_JAVA: _generate_java_project_hierarchy,
     RUNTIME_NODEJS: _generate_nodejs_project_hierarchy,
     RUNTIME_PYTHON: _generate_python_project_hierarchy,
+    RUNTIME_DOTNET: _generate_dotnet_project_hierarchy
 }